+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/ContainerExecDecoratorTest.java
@@ -37,7 +37,7 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.io.output.TeeOutputStream;
 import org.apache.commons.lang.RandomStringUtils;
-import org.junit.AfterClass;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -65,18 +65,15 @@ public class ContainerExecDecoratorTest {
     private ContainerExecDecorator decorator;
 
     @BeforeClass
-    public static void configureCloud() throws Exception {
-        client = setupCloud().connect();
-        deletePods(client, labels, false);
-    }
-
-    @AfterClass
-    public static void after() throws Exception {
-        deletePods(client, labels, false);
+    public static void isKubernetesConfigured() throws Exception {
+        assumeKubernetes();
     }
 
     @Before
-    public void containerExecDecorator() throws Exception {
+    public void configureCloud() throws Exception {
+        client = setupCloud().connect();
+        deletePods(client, labels, false);
+
         String image = "busybox";
         Container c = new ContainerBuilder().withName(image).withImagePullPolicy("IfNotPresent").withImage(image)
                 .withCommand("cat").withTty(true).build();
@@ -87,6 +84,11 @@ public class ContainerExecDecoratorTest {
         System.out.println("Created pod: " + pod.getMetadata().getName());
 
         decorator = new ContainerExecDecorator(client, pod.getMetadata().getName(), image, client.getNamespace());
+    }
+
+    @After
+    public void after() throws Exception {
+        deletePods(client, labels, false);
     }
 
     @Test(timeout = 10000)